refactor: remove obsolete daemon http handlers
